- extand IDriver interface by adding WaitPageLoading methods and implement them - refactored code by adding PageAttribute and apply this functionality to existing code (in Scenarios) - expand PageFactory class - create ReflectionHelper class and implement necessary methods - add LogOut method in BasePage class - add UrlBuilder helper which provide to use relative URL in the Framework and rewrite appropriate code - fix bugs

+++ RozetkaTesting.WebPages/PageComponents/PriceFilterComponent.cs
@@ -103,7 +103,7 @@ namespace RozetkaTesting.WebPages.PageComponents
             TextField_MinPriceValue().ClearAndType("_");
             TextField_MaxPriceValue().ClearAndType("");
             minValue = int.Parse(TextField_MinPriceValue().GetValue());
-            maxValue = int.Parse(TextField_MaxPriceValue().GetValue());
+            maxValue = (int) double.Parse(TextField_MaxPriceValue().GetValue());
         }
 
         private void GetRandomRange(out int minValue, out int maxValue)
